+++ xbudget/__init__.py
@@ -2,7 +2,7 @@
 from .presets import *
 from .collect import *
 from .parse import parse_budgets, BudgetParseError
-from .evaluate import evaluate_budgets
+from .evaluate import evaluate_budgets, MissingDiagnosticError
 from .query import BudgetQuery
 from .display import show_recipe
 from .version import __version__